requirements.txt
azure-storage-blob==12.19.1
requests==2.31.0
SQLAlchemy==2.0.30
Flask==3.0.0
